Change: moved monitor config dbus to checkbox support
